pyproject.toml
[tool.poetry]
name = "flaskonlineshop"
version = "0.1.0"
description = ""
authors = ["Olga Ivanova <olga.ivanova@intexsoft.by>"]

[tool.poetry.dependencies]
python = "3.9.*"
gunicorn = "^20.1.0"
watchdog = "^2.1.6"
Flask-SQLAlchemy = "^2.5.1"
celery = "^5.1.2"
gevent = "^21.8.0"
PyYAML = "^6.0"
argh = "^0.26.2"
redis = "^3.5.3"
SQLAlchemy-Utils = "^0.37.8"
psycopg2-binary = "^2.9.1"
pydantic = "^1.8.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
